refactored symbols and helpers in tokens
