+++ src/tokens/_shared/index.js
@@ -1,0 +1,7 @@
+const helpers = require('./helpers')
+const sharedSymbols = require('./symbols')
+
+module.exports = {
+  ...helpers,
+  sharedSymbols,
+}
