+++ src/tokens/apex/index.js
@@ -1,5 +1,5 @@
 'use strict'
-const { Identifier, createKeywordToken, WhiteSpace } = require('../_shared')
+const { Identifier, createKeywordToken, WhiteSpace, sharedSymbols } = require('../_shared')
 const commentTokens = require('./comments')
 const typeTokens = require('./types')
 const accessLevelTokens = require('./access')
@@ -7,7 +7,6 @@ const loopsAndConditionTokens = require('./loopsAndConds')
 const exceptionTokens = require('./exceptions')
 const classAndMethodTokens = require('./classAndMethods')
 const constantsTokens = require('./constants')
-const sharedSymbolsTokens = require('../_shared/symbols')
 const symbolsTokens = require('./symbols')
 
 const Get = createKeywordToken({
@@ -46,5 +45,5 @@ module.exports = {
   // The Identifier must appear after the keywords because all keywords are valid identifiers.
   Identifier,
   ...symbolsTokens,
-  ...sharedSymbolsTokens,
+  ...sharedSymbols,
 }
